api/requirements.txt
langchain
langchain_community
langchain-text-splitters
langchain-openai
python-dotenv
flask
flask-cors
transformers
torch
tensorflow
tf-keras
deepface
opencv-python
numpy
librosa
pyttsx3
pymongo
cryptography
SpeechRecognition
pydub
openai-whisper